+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/SystemEngineSession.kt
@@ -81,7 +81,6 @@ class SystemEngineSession(
         parent: EngineSession?,
         flags: LoadUrlFlags,
         additionalHeaders: Map<String, String>?,
-        originalInput: String?,
     ) {
         notifyObservers { onLoadUrl() }