+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/TabSessionState.kt
@@ -32,8 +32,6 @@ import java.util.UUID
  * @property lastMediaAccessState - [LastMediaAccessState] detailing the tab state when media started playing.
  * Requires [LastMediaAccessMiddleware] to update the value when playback starts.
  * @property restored Indicates if this page was restored from a persisted state.
- * @property originalInput If the user entered a URL, this is the original user
- * input before any fixups were applied to it.
  */
 data class TabSessionState(
     override val id: String = UUID.randomUUID().toString(),
@@ -47,7 +45,6 @@ data class TabSessionState(
     override val contextId: String? = null,
     override val source: SessionState.Source = SessionState.Source.Internal.None,
     override val restored: Boolean = false,
-    override val originalInput: String? = null,
     val parentId: String? = null,
     val lastAccess: Long = 0L,
     val createdAt: Long = System.currentTimeMillis(),
@@ -110,7 +107,6 @@ fun createTab(
     desktopMode: Boolean = false,
     previewImageUrl: String? = null,
     hasFormData: Boolean = false,
-    originalInput: String? = null,
 ): TabSessionState {
     return TabSessionState(
         id = id,
@@ -143,7 +139,6 @@ fun createTab(
         ),
         mediaSessionState = mediaSessionState,
         historyMetadata = historyMetadata,
-        originalInput = originalInput,
     )
 }