+++ mobile/android/android-components/components/feature/session/src/main/java/mozilla/components/feature/session/SessionUseCases.kt
@@ -37,7 +37,6 @@ class SessionUseCases(
             url: String,
             flags: LoadUrlFlags = LoadUrlFlags.none(),
             additionalHeaders: Map<String, String>? = null,
-            originalInput: String? = null,
         )
     }
 
@@ -54,16 +53,13 @@ class SessionUseCases(
          * @param url The URL to be loaded using the selected session.
          * @param flags The [LoadUrlFlags] to use when loading the provided url.
          * @param additionalHeaders the extra headers to use when loading the provided url.
-         * @param originalInput If the user entered a URL, this is the
-         * original user input before any fixups were applied to it.
          */
         override operator fun invoke(
             url: String,
             flags: LoadUrlFlags,
             additionalHeaders: Map<String, String>?,
-            originalInput: String?,
         ) {
-            this.invoke(url, store.state.selectedTabId, flags, additionalHeaders, originalInput)
+            this.invoke(url, store.state.selectedTabId, flags, additionalHeaders)
         }
 
         /**
@@ -75,15 +71,12 @@ class SessionUseCases(
          * @param sessionId the ID of the session for which the URL should be loaded.
          * @param flags The [LoadUrlFlags] to use when loading the provided url.
          * @param additionalHeaders the extra headers to use when loading the provided url.
-         * @param originalInput If the user entered a URL, this is the
-         * original user input before any fixups were applied to it.
          */
         operator fun invoke(
             url: String,
             sessionId: String? = null,
             flags: LoadUrlFlags = LoadUrlFlags.none(),
             additionalHeaders: Map<String, String>? = null,
-            originalInput: String? = null,
         ) {
             val loadSessionId = sessionId
                 ?: store.state.selectedTabId
@@ -99,7 +92,6 @@ class SessionUseCases(
                     url = url,
                     flags = flags,
                     additionalHeaders = additionalHeaders,
-                    originalInput = originalInput,
                 )
                 store.dispatch(
                     EngineAction.OptimizedLoadUrlTriggeredAction(