+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/HomeActivity.kt
@@ -1231,13 +1231,11 @@ open class HomeActivity : LocaleAwareAppCompatActivity(), NavHostActivity {
                     flags = flags,
                     private = private,
                     historyMetadata = historyMetadata,
-                    originalInput = searchTermOrURL,
                 )
             } else {
                 components.useCases.sessionUseCases.loadUrl(
                     url = searchTermOrURL.toNormalizedUrl(),
                     flags = flags,
-                    originalInput = searchTermOrURL,
                 )
             }
         } else {