+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/search/awesomebar/AwesomeBarView.kt
@@ -73,7 +73,6 @@ class AwesomeBarView(
             url: String,
             flags: EngineSession.LoadUrlFlags,
             additionalHeaders: Map<String, String>?,
-            originalInput: String?,
         ) {
             interactor.onUrlTapped(url, flags)
         }