+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/webcompat/fake/FakeEngineSession.kt
@@ -35,7 +35,6 @@ internal class FakeEngineSession(
         parent: EngineSession?,
         flags: LoadUrlFlags,
         additionalHeaders: Map<String, String>?,
-        originalInput: String?,
     ) {}
 
     override fun loadData(data: String, mimeType: String, encoding: String) {}